app/pyproject.toml
[project]
name = "nubrain"
version = "0.0.5"
description = "NuBrain connect"
authors = [{ name = "nuBrain Team", email = "hello@nubrain.io" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "brainflow>=5.16.0",
    "certifi>=2025.4.26",
    "charset-normalizer>=3.4.2",
    "h5py",
    "idna>=3.10",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
    "PyYAML",
    "requests>=2.32.3",
    "scipy>=1.15.3",
    # "setuptools>=80.9.0",        # Needed?
    "urllib3>=2.4.0",
]
license = "MIT"
license-files = ["LICENSE"]

# Development dependencies.
[project.optional-dependencies]
dev = ["build", "spyder-kernels==2.5.2", "twine"]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nubrain"]

[project.urls]
Homepage = "https://nubrain.io/"

[project.scripts]
nubrain = "nubrain.cli:main"
